Remove env specific test files for universal tests
